Add files via upload
Admin panel of the quiz mobile app.
This code enables the admin panel to fetch, send, delete and update data in real-time from the database. The admin panel perform various tasks such as adding new quiz questions, updating user information, and monitoring the performance of the app.
Using this code, the administrator can fetch data from the database and display it in the admin panel, allowing for easy management of the quiz questions, users, and other data. The code also allows the administrator to update the data in real-time, which means that any changes made to the database are immediately reflected in the admin panel. In addition, the code also enables the administrator to delete data from the database. This feature is particularly useful in cases where certain quiz questions or user information need to be removed from the system. 
Overall, this code provides a powerful set of tools for managing the quiz mobile app, allowing the administrator to make real-time updates and changes to the data in the database.